EDSC-3750: Addressing PR feedback
