--- static/src/js/components/GranuleImage/GranuleImage.scss
@@ -1,95 +1,0 @@
-.granule-image {
-  position: absolute;
-  top: 100px;
-  right: 1.5rem;
-  width: 0;
-  height: 0;
-  z-index: 401; /* Sits above .leaflet-map-pane at index 400 */
-  color: $color__white;
-  background-color: $color__black;
-  transition: all 0.2s ease;
-
-  &--is-open {
-    width: 16rem;
-    height: auto;
-  }
-
-  &__container {
-    width: 100%;
-    height: auto;
-    width: 16rem;
-    max-height: 16rem;
-    padding: 0.5rem;
-    box-shadow: 0 0 1rem rgba($color__black--900, 0.35);
-    transform-origin: top right;
-    transform: scale(0);
-    transition: transform 0.2s ease-in-out;
-    border: 2px solid rgba(0,0,0,.2);
-    overflow: hidden;
-
-    .granule-image--is-open & {
-      transform: scale(1);
-    }
-  }
-
-  &__image {
-    width: 100%;
-    height: auto;
-    max-height: 15rem;
-    opacity: 0;
-    object-fit: contain;
-
-    &--is-loaded {
-      opacity: 1;
-    }
-  }
-
-  &__button {
-    position: absolute;
-    top: -0.925rem;
-    right: -0.925rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0;
-    width: 1.75rem;
-    height: 1.75rem;
-    border-radius: 50%;
-    line-height: 1.75rem;
-    color: $color__white;
-    box-shadow: 0 0 1rem rgba($color__black--900, 0.35);
-    z-index: 1;
-
-    &:hover,
-    &:active {
-      color: $color__white;
-    }
-
-    &--close {
-      background-color: $color__red;
-    }
-
-    &--open {
-      background-color: $color__green;
-
-      .button__icon {
-        top: 0.05rem;
-        position: relative;
-      }
-    }
-  }
-
-  &__placeholder {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: $color__black--700;
-    width: 16rem;
-    min-height: 16rem;
-  }
-}
