+++ static/src/js/containers/MapContainer/MapWrapper.js
@@ -10,7 +10,6 @@ import crsProjections from '../../util/map/crs'
 
 import ConnectedSpatialSelectionContainer from '../SpatialSelectionContainer/SpatialSelectionContainer'
 import GranuleGridLayer from '../../components/Map/GranuleGridLayer'
-// import GranuleImageContainer from '../GranuleImageContainer/GranuleImageContainer'
 import LayerBuilder from '../../components/Map/LayerBuilder'
 import MouseEventsLayer from '../../components/Map/MouseEventsLayer'
 import ProjectionSwitcher from '../../components/Map/ProjectionSwitcher'
